+++ src/cli/data.rs
@@ -5,20 +5,20 @@ use nu_errors::ShellError;
 use nu_protocol::{Signature, UntaggedValue};
 use nu_source::Tag;
 
-pub struct Kv;
+pub struct Data;
 
 #[async_trait]
-impl nu_cli::WholeStreamCommand for Kv {
+impl nu_cli::WholeStreamCommand for Data {
     fn name(&self) -> &str {
-        "kv"
+        "data"
     }
 
     fn signature(&self) -> Signature {
-        Signature::build("kv")
+        Signature::build("data")
     }
 
     fn usage(&self) -> &str {
-        "Perform Key/Value operations against a bucket"
+        "Performs operations against the data service"
     }
 
     async fn run(
@@ -28,7 +28,7 @@ impl nu_cli::WholeStreamCommand for Kv {
     ) -> Result<OutputStream, ShellError> {
         let registry = registry.clone();
         let stream = stream! {
-            yield UntaggedValue::string(nu_cli::get_help(&Kv, &registry))
+            yield UntaggedValue::string(nu_cli::get_help(&Data, &registry))
             .into_value(Tag::unknown())
         };
         Ok(OutputStream::from_input(stream))