+++ src/cli/doc_remove.rs
@@ -1,38 +1,38 @@
-//! The `kv-replace` command performs a KV replace operation.
-
-use super::util::{json_rows_from_input_columns, json_rows_from_input_optionals};
+//! The `kv remove` command performs a KV remove operation.
 
 use crate::state::State;
-use couchbase::ReplaceOptions;
+use couchbase::RemoveOptions;
 
 use async_trait::async_trait;
+use futures::stream::StreamExt;
 use log::debug;
 use nu_cli::{CommandArgs, CommandRegistry, OutputStream};
 use nu_errors::ShellError;
-use nu_protocol::{Signature, SyntaxShape, TaggedDictBuilder, UntaggedValue};
+use nu_protocol::{
+    MaybeOwned, Primitive, Signature, SyntaxShape, TaggedDictBuilder, UntaggedValue,
+};
 use nu_source::Tag;
 use std::sync::Arc;
 
-pub struct KvReplace {
+pub struct DocRemove {
     state: Arc<State>,
 }
 
-impl KvReplace {
+impl DocRemove {
     pub fn new(state: Arc<State>) -> Self {
         Self { state }
     }
 }
 
 #[async_trait]
-impl nu_cli::WholeStreamCommand for KvReplace {
+impl nu_cli::WholeStreamCommand for DocRemove {
     fn name(&self) -> &str {
-        "kv replace"
+        "doc remove"
     }
 
     fn signature(&self) -> Signature {
-        Signature::build("kv replace")
+        Signature::build("doc remove")
             .optional("id", SyntaxShape::String, "the document id")
-            .optional("content", SyntaxShape::String, "the document content")
             .named(
                 "id-column",
                 SyntaxShape::String,
@@ -45,16 +45,10 @@ impl nu_cli::WholeStreamCommand for KvReplace {
                 "the name of the bucket",
                 None,
             )
-            .named(
-                "content-column",
-                SyntaxShape::String,
-                "the name of the content column if used with an input stream",
-                None,
-            )
     }
 
     fn usage(&self) -> &str {
-        "Replace a document through Key/Value"
+        "Removes a document through the data service"
     }
 
     async fn run(
@@ -62,16 +56,21 @@ impl nu_cli::WholeStreamCommand for KvReplace {
         args: CommandArgs,
         registry: &CommandRegistry,
     ) -> Result<OutputStream, ShellError> {
-        run_replace(self.state.clone(), args, registry).await
+        run_get(self.state.clone(), args, registry).await
     }
 }
 
-async fn run_replace(
+async fn run_get(
     state: Arc<State>,
     args: CommandArgs,
     registry: &CommandRegistry,
 ) -> Result<OutputStream, ShellError> {
     let mut args = args.evaluate_once(registry).await?;
+
+    let id_column = args
+        .get("id-column")
+        .map(|id| id.as_string().unwrap())
+        .unwrap_or_else(|| String::from("id"));
 
     let bucket_name = match args
         .get("bucket")
@@ -86,40 +85,48 @@ async fn run_replace(
         }
     };
 
-    let id_column = args
-        .get("id-column")
-        .map(|id| id.as_string().unwrap())
-        .unwrap_or_else(|| String::from("id"));
+    let mut ids = vec![];
+    while let Some(item) = args.input.next().await {
+        let untagged = item.into();
+        match untagged {
+            UntaggedValue::Primitive(p) => match p {
+                Primitive::String(s) => ids.push(s.clone()),
+                _ => {}
+            },
+            UntaggedValue::Row(d) => {
+                if let MaybeOwned::Borrowed(d) = d.get_data(id_column.as_ref()) {
+                    let untagged = &d.value;
+                    if let UntaggedValue::Primitive(p) = untagged {
+                        if let Primitive::String(s) = p {
+                            ids.push(s.clone())
+                        }
+                    }
+                }
+            }
+            _ => {}
+        }
+    }
 
-    let content_column = args
-        .get("content-column")
-        .map(|content| content.as_string().unwrap())
-        .unwrap_or_else(|| String::from("content"));
-
-    let mut rows = json_rows_from_input_columns(&mut args, &id_column, &content_column).await?;
-    rows.extend(json_rows_from_input_optionals(&mut args)?);
+    if let Some(id) = args.nth(0) {
+        ids.push(id.as_string()?);
+    }
 
     let bucket = state.active_cluster().bucket(&bucket_name);
     let collection = bucket.default_collection();
 
-    debug!("Running kv replace for docs {:?}", &rows);
+    debug!("Running kv remove for docs {:?}", &ids);
 
     let mut results = vec![];
-    for (id, content) in rows.iter() {
-        match collection
-            .replace(id, content, ReplaceOptions::default())
-            .await
-        {
+    for id in ids {
+        match collection.remove(&id, RemoveOptions::default()).await {
             Ok(res) => {
                 let tag = Tag::default();
                 let mut collected = TaggedDictBuilder::new(&tag);
-                collected.insert_value(&id_column, id.clone());
+                collected.insert_value(&id_column, id);
                 collected.insert_value("cas", UntaggedValue::int(res.cas()).into_untagged_value());
                 results.push(collected.into_value());
             }
-            Err(e) => {
-                return Err(ShellError::untagged_runtime_error(format!("{}", e)));
-            }
+            Err(_e) => {}
         };
     }
     Ok(OutputStream::from(results))